Fuck OpenStack
Add in support to grab existing rules from the OpenStack distro, then
essentially reduce them into one liners so we can override the existing
policies.  Then, because OpenStack is essentially bollocks, we need to
work around its inadequacies to make it even work, because half the
infromation required is missing, and a lot of what would be required is
actually hard coded to fail.